refactor: DAO 클래스 병합
- 테이블 당 하나의 DAO를 유지하기 위함
